Improve FLBasicSerializationTest>>#testWideStringClassName
This test change the name of a class after creating it but this seems to create a problem that this class ends up been an obsolete class once we cleanup after the test.

To not let any leftover I propose to create it directly with the right name from the start. With this, no obsolete class stays in the system afterward