Fix bug with obsolete classes and extension methods
If you recompile a class with extension methods and remove it afterward, then the class stays as obsolete. This happens because of a bug creating duplicated entries in the packages extension maps. 

I'm proposing a fix to not add new entries to the extension maps in case of recompilation. Coming with a comment and a regression test.

Fixes #17927