Fix the logic of the extra commands on methods
In Calypso Hernan started to migrate commands from Commander1 to Commander2. This brought a change in the logic but the commands in the extra group on methods lost the ability to execute the #prepareFullExecution method which cause some commands to fail.

I'm fixing this + I renamed a class to have a better name

Fixes #17506