Remove ExternalProjectLoader
This class seems to be dead code? Probably used in the past for the bootstrap?

The class comment says it's to have versions tested with Pharo but the projects present inside do not seems to match. For example Roassal is in Pharo already and the repo in the #allProjects method is not the official repo anymore