Improve packaging of Ston
We currently have a package `STON-Text support` containing methods to export and import Text in Ston. But we also have STON-Extensions that add extensions to multiple classes in Pharo to customize their import/export so I do not see why text support should be a special case. 
On top of that, most extensions are done directly in STON-Core so I'm not even sure we really need this STON-Extensions package. But this change focuses on 'STON-Text support'. 

Fixes #4505