Fix #isDeprecated logic of CompiledMethod
When a class is in a deprecated package, we consider it as deprecated.
But if a method is in a deprecated package or class, it is not considered deprecated. 

To me this mixed behavior is not good. After talking with people, they seems to expect the deprecation to be passed on so I'm updating CompileMethod>>#isDeprecated to take the class and package state into account for deprecations.