Improve the printing of errors on executable comments
Currently if we have a syntax error on an executable comment we have 0 infromation on which comment it is. We just get `Doc comment is not well formed` as a log and `PharoDocComment.Tests.CommentTestCase.nil` as the test been executed.

I propose to add more info to the log to know the class, selector and content of the comment.