Fix equality between MCClassDefinition and MCTraitDefinition
Currently if you try to compare a TraitDefinition with a ClassDefinition you will get a DNU. I'm fixing it by stopping the comparison if the classes are not the same.

This will allow us to commit the change of a class to a trait of the same name. Currently we cannot commit with Iceberg this kind of changes 

Fixes #18035