Remove dependency of Network-Mail to Morphic
MailMessage has some dependencies over Morphic and UIManager to be able to view a MailMessage content. 

I propose to move those methods to NewTools and to use Spec instead of Morphic to view the content. 

I already did a PR to NewTools so we can just remove those two methods from Pharo.

https://github.com/pharo-spec/NewTools/pull/1104