RGMethodDefinitionTest>>testAsFullMethodDefinition should work from the class's full ring definition.
Previously #testAsFullMethodDefinition retrieved the method's full ring definition and then assumed that the class contained the full ring definition.  This relied on the incorrect behaviour of RGMethodDefinition>>realParent setting the model parent.  It also clearly contradicts the comment in CompiledMethod>>asFullRingDefinition which states that "the full conversion does not happen at the level of the class. If you need that request asFullRingDefinition to the class".

fix: #18444