Fix error report generation during the bootstrap
Currently if we end up in an error before loading the System-Time package, we have the error report that is crashing because it is trying to print the date and time.
This change makes it optional to print the time before the classes Date and Time and loaded in the image. 
This allows is to at least have an minimal error report during the bootstrap.