OpalCompiler should not depend on Roassal
The compiler is using SequenceableCollection>>#min:/max: but those methods are from Roassal. 

I propose to move them directly on SequenceableCollection

I ended up in this code in the bootstrap. So we either need to make the compiler not depend on those methods or to move them in the original class. I propose the later but I am also open to the first