Simplify the registering of OpalCompiler
Up until now we have SmalltalkImage looking if Opal is in the system and we also register in the bootstrap the compiler.

I propose to simplify things. SmalltalkImage should not know the compiler. And I removed the explicit registration from the bootstrap. Instead, when we load Opal, it will check if a compiler is registered. If not, it will register itself.

This is a small cleaning to remove a bad dependency and simplify (a tiny bit) the bootstrap