Remove Images-Animated
This package contains 2 classes. One morph that is not used. And one that is now only used by the GIFReadWritter.

I propose to move the morph the Deprecated-Morphic and the AnimatedImageFrame to coexist with the GIFReadWritter.

This cuts the dependency of Graphics-Files to Morphic! We can now have Image readers and writters without having Morphic