Move Deprecations and co out of the Kernel
Deprecation, Automatic rewriting and BackwardCompatibility are classes relying on the compiler to produce transformations but the Kernel should not depend on the compiler.
Since they cannot work before we load the compiler, I propose to move all those classes out of the Kernel in their own package and to load them alongside the compiler.