Bootstrap: Remove some duplicated load spots
Packages to load for the core of monticello needs to be declared in 2 places (1 outside of the loaded Pharo code).

Packages to load for remote repositories of Monticello needs to be declared in 3 places (2 outside of the loaded Pharo code).

With this change, we will only need to declare them in one place