Cut dependency of Collection-Strings to Unicode
Currently Collections-Strings depends explicitly on Unicode. It would lead to error to move the methods as extension of Unicode because the error would not be clear.
I propose to make the dependency  weaker by checking if unicode is in the image and raise a right warning if it is not in the image.