Cut multiple dependencies of Command line handler
This changes is cutting multiple dependencie of PharoCommandLineHandler to some projects:
- FileSystem-Core
- UIManager
- StartupPreferences

In order to do this, I splited PharoOptionHandler into two. BasicPharoOptionHandler that has the minimal options handling and PharoOtpionHandler loaded later that is managing the preferences.

I also cleaned some baselines.

In a future step I also plan to cut also the dependency to System-Hashing but that will be in a future step because I will also add back the password manager with this step.