Start SessionManager with a session
Initialize SessionManager instances with a default session instead of calling it explicitly.

This seems like a small change but the ErrorHandler needs the SessionManager to be initialized to log errors. So if an error happens before the end of the kernel initialization, the error handler cannot work if the SessionManager has no session.
With this, we will be able to have more error report