Extract SystemVersion from System-Support
System-Support needs to be in the core of Pharo. But SystemVersion depends on System-Time which is not part of the core of Pharo.
We do not need the version until it is initialized in the bootstrap so I propose to create a package System-Version to load later in the bootstrap. This will reduce the number of unwanted dependencies in the core of the bootstrap.

I also removed some other unwanted dependencies.

Little note if people have an opinion: Currently we have the version of multiple things such the OS, the VM, the Pharo version...
I wonder if the OS and System version should not be accessible via the SystemVersion instead of the Smalltalk global? But it would be for another PR