Fuel adds a variant #nextBytesPutAll:  of nextPutAll:
The only change is that it relaxes the guard and instead of testing the class. it tests the class layout:

	collection class instSpec == aCollection class instSpec ifFalse:
 		[^ super nextPutAll: aCollection ].

as the follwowing code sends a method to collection, this should be as save to do as the class check that was there before.

Why does fuel do it? It allows the serializer to write all VariableByte instances directy to a stream on a ByteArray,  without having to create a copy.
(this method is used by Soil, too)

The PR changes nextPutAll: and deprecates nextBytesPutAll:

Both Fuel and Stream/Collection tests are green