Remove the concept of ignored dependencies
In the image we have two ways of declaring that we are aware we have a bad dependency. The first is to add the package name to #ignoredDependencies. The second is to add exceptions in SystemDependencyTest.
This makes it harder to find the dependencies to remove (I did not notive #ignoredDependencies for a long time).

I found that most of the ignored dependencies are outdated and I could fix the remaining ones. So I propose to keep only one "whitelist" system instead of 2.

Changes:
- Remove ignored dependencies
- Remove some dead code left over ofter the MCZ export removal
- Remove a dead method in the compiler calling an unimplemented method
- Remove OCIRReconstructor that is a deprecated and even got renamed