Speed up MCClassDefinition creation and clean MCMethodDefinition
This change brings two improvements to Monticello.

The first is the simplification of MCMethodDefinition. I added back a cache that was present in the past with a fix. But I also simplified a piece of Monticello code reccently making this cache useless because now it takes the same time to check the cache and to create the method definition. So it's better to use the method with the simpler code.

I also speed up the creation of MCClassDefinition and MCTraitDefinition by making the retrieval of the comment stamp lazy because Iceberg does not need this comment stamp and it is the bigger user of Monticello model currently.