Metacello: Deprecate configuration API
This change is deprecating some of the API that was specific to ConfigurationOf since we have the matadatas of project in git for BaselineOf such as:
- author
- blessing
- timestamp
- description
- configuration:with:
- versionString 

I removed all the code related to timestamp, author, blessing and description. I still need to remove a lot of code related to the version string but this code is deeply rooted in Metacello so I'll need to do that in multiple steps to ensure I do not break anything. 
I also thing we can remove the #operator API but it is too deeply rooted in Metacello to be sure for now. I'll do that later also