Metacello: clean MetacelloProjectSpec hierarchy
This is the start of a pretty big refactoring (or series of refactorings) I want to do in the model of Metacello to simpify it!


The hierarchy of MetacelloProjectSpec is here to manage the declaration of dependencies. In the past this hierarchy was like this:

- MetacelloProjectSpec (abstract)
-- MetacelloBaselineOfProjectSpec : defines a dependency via another baseline
-- MetacelloConfigurationOfProjectSpec : defines a dependency via another configuration
-- MetacelloNamelessProjectSpec : defines a dependency as a nameless "project" that will be cast as a configuration or baseline project spec before the actual actions of Metacello

Now we do not have the MetacelloConfigurationOfProjectSpec anymore and MetacelloNamelessProjectSpec is almost empty because all its behavior is in MetacelloProjectSpec.


This change is pushing all the behavior specific to MetacelloNamelessProjectSpec into this class instead of having the behavior in the abstract superclass. This will allow to have a clearer understanding of what are the differences between a NamelessProjectSpec and a BaselineOfProjectSpec.


I have two future steps in mind.

The first future step is to kill the concept of multiple repositories. With configurations, we could look for the project in multiple repositories. But baselines are storing more info in the "repository" since it points for example to branches or versions. Because of that, MetacelloBaselineOfProjectSpec is limited to only one repository. And since MetacelloNamelessProjectSpec **must** be cast into a MetacelloBaselineOfProjectSpec, only the last defined repository will win in the end. 

I want to refactore Metacello API to deprecate this system of multiple repositories and allow only one since it will be the last one (or first one in some cases? it is still not clear to me) that will win. This should simplify further the model of Metacello!

Second, once this change is done, most of the difference between MetacelloNamelessProjectSpec and MetacelloBaselineOfProjectSpec will be gone. I would like to remove entirely the concept of "project dependency" to keep only "baseline of project spec" kind of dependencies. I would rename it "MetacelloDependencySpec". This will remove the need of useless casts. This will simplifcy the API. This will simplify the model. And I believe that we can find much more deadcode if we do that because I am currently unsure if some code is dead or not because of those casts >w<