Improve error while asking the parse tree of a method without source
I recently got 2 PR failing the release tests because not all methods have sources. I doubt the problem comes from my PRs but it is hard to find where it comes from. 
I'm proposing to at least print the name of the method provoking the problem when this happens so that we can find the origin of the bug in an easier way