Metacello: improve management of deprecated API
Rome of the API that has been deprecated in Pharo 14 of Metacello needs to stay. If we remove it, then Baselines managing projects that can load in other smalltalks or in previous versions of Pharo will not be able to load anymore.
I replaced the errors/deprecations with warnings in the transcript. 

This fixed the loading of Seaside. Fixes #19091